requirements.txt
lxml>=0
MechanicalSoup>=0
parse>=0
pysocks>=0
python-twitter>=0
requests>=0
